Separate info control into separate file because it might be more complex in future.
